chore: extract load-broadcast-from-db suite into src/suites/load-broadcast-from-db.ts
Part of REAL-1037.
